+++ df-chat-enhance/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Release 5.0.2 (2025-03-12)
 - **FIX #513:** UserConfig properly saves its data again.
+- **FIX #512:** Inline-Rolls containing markdown characters in chat function correctly again without being converted to html.
 
 ## Release 5.0.1 (2025-03-01)
 - **UPDATE #468:** Update Chinese Localization: Courtesy of [LiyuNodream](https://github.com/LiyuNodream).